refactor(machine): decompose expander_time_continuation.go into 4 files
Extract let-syntax/letrec-syntax (expander_let_syntax.go), primitive form
handlers (expander_primitive_forms.go), and lambda/case-lambda expansion
(expander_lambda.go) from expander_time_continuation.go.

Pure file reorganization — no behavior changes.
