+++ machine/expander_time_continuation.go
@@ -40,7 +40,6 @@ package machine
 
 import (
 	"context"
-	"strings"
 
 	"github.com/aalpar/wile/environment"
 	"github.com/aalpar/wile/internal/syntax"
@@ -169,863 +168,6 @@ func (p *ExpanderTimeContinuation) ExpandPrimitiveForm(primName string, sym *syn
 	}
 	// Unknown primitive - return unchanged (safe default)
 	return syntax.NewSyntaxCons(sym, expr, sym.SourceContext()), nil
-}
-
-// expandUnchanged returns the form unchanged. Used for forms whose subexpressions
-// are processed at a later stage (compile time) or should not be expanded (quote).
-//
-// Forms using this expander:
-//   - quote: Content is literal data, not code
-//   - define-syntax: Transformer is compiled, not expanded
-//   - quasiquote: Has special expansion rules handled at compile time
-//   - unquote, unquote-splicing: Only valid inside quasiquote
-//   - include, include-ci: Files are read at compile time
-//   - cond-expand: Feature expressions use special syntax, not macros
-//   - syntax, syntax-case, quasisyntax, unsyntax, unsyntax-splicing, with-syntax:
-//     Compile-time forms handled during compilation
-func (p *ExpanderTimeContinuation) expandUnchanged(sym *syntax.SyntaxSymbol, expr syntax.SyntaxValue) (syntax.SyntaxValue, error) {
-	return syntax.NewSyntaxCons(sym, expr, sym.SourceContext()), nil
-}
-
-// expandLetSyntax fully expands let-syntax during the expansion phase.
-// Creates local macro bindings, expands the body, and returns the expanded result.
-// The let-syntax wrapper disappears - only the expanded body remains.
-//
-// R7RS §4.3.1: let-syntax establishes local macro definitions visible only in the body.
-func (p *ExpanderTimeContinuation) expandLetSyntax(sym *syntax.SyntaxSymbol, expr syntax.SyntaxValue) (syntax.SyntaxValue, error) {
-	return p.expandLetSyntaxImpl(sym, expr, false)
-}
-
-// expandLetrecSyntax fully expands letrec-syntax during the expansion phase.
-// Like let-syntax but transformers can reference each other (mutual recursion).
-//
-// R7RS §4.3.1: letrec-syntax is like let-syntax but with mutual visibility.
-func (p *ExpanderTimeContinuation) expandLetrecSyntax(sym *syntax.SyntaxSymbol, expr syntax.SyntaxValue) (syntax.SyntaxValue, error) {
-	return p.expandLetSyntaxImpl(sym, expr, true)
-}
-
-// expandLetSyntaxImpl implements both let-syntax and letrec-syntax expansion.
-// The recursive parameter controls whether bindings can see each other.
-//
-// This function:
-// 1. Creates a child expand environment with local macro bindings
-// 2. Compiles each syntax-rules transformer
-// 3. Expands body expressions with the child environment
-// 4. Wraps in lambda if body contains defines (for scope isolation)
-// 5. Returns the expanded body - the let-syntax wrapper disappears
-func (p *ExpanderTimeContinuation) expandLetSyntaxImpl(sym *syntax.SyntaxSymbol, expr syntax.SyntaxValue, recursive bool) (syntax.SyntaxValue, error) {
-	formName := "let-syntax"
-	if recursive {
-		formName = "letrec-syntax"
-	}
-	sc := sym.SourceContext()
-
-	// expr is (<bindings> <body>) - args after keyword
-	argsPair, ok := expr.(*syntax.SyntaxPair)
-	if !ok || argsPair.IsEmptyList() {
-		return nil, werr.WrapForeignErrorf(werr.ErrInvalidSyntax, "%s: expected bindings and body", formName)
-	}
-
-	// Get the bindings list
-	bindingsStx := argsPair.SyntaxCar()
-	bindingsEmpty := syntax.IsSyntaxEmptyList(bindingsStx)
-	var bindingsPair *syntax.SyntaxPair
-	if !bindingsEmpty {
-		var pairOk bool
-		bindingsPair, pairOk = bindingsStx.(*syntax.SyntaxPair)
-		if !pairOk {
-			return nil, werr.WrapForeignErrorf(werr.ErrInvalidSyntax, "%s: expected bindings list", formName)
-		}
-	}
-
-	// Get the body
-	bodyStx := argsPair.SyntaxCdr()
-	bodyPair, ok := bodyStx.(*syntax.SyntaxPair)
-	if !ok || bodyPair.IsEmptyList() {
-		return nil, werr.WrapForeignErrorf(werr.ErrInvalidSyntax, "%s: expected body expressions", formName)
-	}
-
-	// Count bindings for local environment allocation
-	numBindings := 0
-	var current *syntax.SyntaxPair
-	if !bindingsEmpty {
-		current = bindingsPair
-		for !syntax.IsSyntaxEmptyList(current) {
-			numBindings++
-			cdr := current.SyntaxCdr()
-			nextPair, ok := cdr.(*syntax.SyntaxPair)
-			if ok {
-				current = nextPair
-			} else {
-				break
-			}
-		}
-	}
-
-	// Create child expand environment for macro bindings.
-	// Use p.env directly as the parent (not p.env.Expand()) to preserve
-	// the environment chain for nested let-syntax. When we have:
-	//   (let-syntax ((outer ...))
-	//     (let-syntax ((inner ...)) ...))
-	// The inner let-syntax's environment must have outer's environment
-	// in its parent chain, not the global expand environment.
-	localExpandEnv := environment.NewLocalEnvironment(numBindings)
-	childExpandEnv := environment.NewEnvironmentFrameWithParent(localExpandEnv, p.env)
-
-	// Create a rebinding scope for the let-syntax body.
-	// Rebinding scopes indicate that auxiliary syntax could be shadowed.
-	// This is used in literalScopesMatch to correctly reject shadowed literals.
-	letScope := syntax.NewRebindingScopeWithLabel("let-syntax")
-
-	// For letrec-syntax, pre-register all keywords so transformers can see each other
-	if recursive && !bindingsEmpty {
-		current = bindingsPair
-		for !syntax.IsSyntaxEmptyList(current) {
-			bindingStx := current.SyntaxCar()
-			bindingPair, ok := bindingStx.(*syntax.SyntaxPair)
-			if !ok || bindingPair.IsEmptyList() {
-				return nil, werr.WrapForeignErrorf(werr.ErrInvalidSyntax, "%s: invalid binding", formName)
-			}
-			keywordStx := bindingPair.SyntaxCar()
-			keywordSym, ok := keywordStx.(*syntax.SyntaxSymbol)
-			if !ok {
-				return nil, werr.WrapForeignErrorf(werr.ErrNotASymbol, "%s: keyword must be a symbol", formName)
-			}
-			keyword := keywordSym.Unwrap().(*values.Symbol)
-			// Create binding with letScope so free identifier resolution works
-			_, _ = childExpandEnv.MaybeCreateLocalBindingWithScopes(keyword, environment.BindingTypeSyntax, []*syntax.Scope{letScope}, keywordSym.SourceContext())
-
-			cdr := current.SyntaxCdr()
-			nextPair, ok := cdr.(*syntax.SyntaxPair)
-			if ok {
-				current = nextPair
-			} else {
-				break
-			}
-		}
-	}
-
-	// Compile each transformer and store in child expand environment
-	if !bindingsEmpty {
-		current = bindingsPair
-	}
-	for !bindingsEmpty && !syntax.IsSyntaxEmptyList(current) {
-		bindingStx := current.SyntaxCar()
-		bindingPair, ok := bindingStx.(*syntax.SyntaxPair)
-		if !ok || bindingPair.IsEmptyList() {
-			return nil, werr.WrapForeignErrorf(werr.ErrInvalidSyntax, "%s: invalid binding", formName)
-		}
-
-		// Get keyword
-		keywordStx := bindingPair.SyntaxCar()
-		keywordSym, ok := keywordStx.(*syntax.SyntaxSymbol)
-		if !ok {
-			return nil, werr.WrapForeignErrorf(werr.ErrNotASymbol, "%s: keyword must be a symbol", formName)
-		}
-		keyword := keywordSym.Unwrap().(*values.Symbol)
-
-		// Get transformer expression
-		transformerCdr := bindingPair.SyntaxCdr()
-		transformerPair, ok := transformerCdr.(*syntax.SyntaxPair)
-		if !ok || transformerPair.IsEmptyList() {
-			return nil, werr.WrapForeignErrorf(werr.ErrInvalidSyntax, "%s: missing transformer expression", formName)
-		}
-		transformerExpr := transformerPair.SyntaxCar()
-
-		// Check if transformer is a syntax-rules form
-		transformerPairExpr, ok := transformerExpr.(*syntax.SyntaxPair)
-		if !ok {
-			return nil, werr.WrapForeignErrorf(werr.ErrUnsupportedTransformer, "%s: only syntax-rules transformers are currently supported", formName)
-		}
-		car := transformerPairExpr.SyntaxCar()
-		if car == nil {
-			return nil, werr.WrapForeignErrorf(werr.ErrUnsupportedTransformer, "%s: invalid transformer", formName)
-		}
-		srSym, ok := car.(*syntax.SyntaxSymbol)
-		if !ok {
-			return nil, werr.WrapForeignErrorf(werr.ErrUnsupportedTransformer, "%s: only syntax-rules transformers are currently supported", formName)
-		}
-		srSymVal := srSym.Unwrap()
-		srSymbol, ok := srSymVal.(*values.Symbol)
-		if !ok || srSymbol.Key != "syntax-rules" {
-			return nil, werr.WrapForeignErrorf(werr.ErrUnsupportedTransformer, "%s: only syntax-rules transformers are currently supported", formName)
-		}
-
-		// Compile the syntax-rules transformer
-		closure, err := CompileSyntaxRules(p.ctx, p.env, transformerPairExpr, p.libraryScope)
-		if err != nil {
-			return nil, werr.WrapForeignErrorf(err, "%s: could not compile transformer for %s", formName, keyword.Key)
-		}
-
-		// Store in child expand environment with letScope for free identifier resolution
-		localIndex, created := childExpandEnv.MaybeCreateLocalBindingWithScopes(keyword, environment.BindingTypeSyntax, []*syntax.Scope{letScope}, keywordSym.SourceContext())
-		if !created {
-			localIndex = childExpandEnv.GetLocalIndex(keyword)
-		}
-		if localIndex != nil {
-			err := childExpandEnv.SetLocalValue(localIndex, closure)
-			if err != nil {
-				return nil, err
-			}
-		}
-
-		cdr := current.SyntaxCdr()
-		nextPair, ok := cdr.(*syntax.SyntaxPair)
-		if ok {
-			current = nextPair
-		} else {
-			break
-		}
-	}
-
-	// Add the let-syntax scope to the body
-	scopedBody := bodyPair.AddScope(letScope)
-	scopedBodyPair, ok := scopedBody.(*syntax.SyntaxPair)
-	if !ok {
-		return nil, werr.WrapForeignErrorf(werr.ErrNotAList, "%s: body must be a list", formName)
-	}
-
-	// Create expander with child expand environment for body expansion
-	childExpander := NewExpanderTimeContinuation(p.ctx, childExpandEnv)
-
-	// Expand all body expressions and check for defines
-	var expandedExprs []syntax.SyntaxValue
-	hasDefine := false
-	current = scopedBodyPair
-	for !syntax.IsSyntaxEmptyList(current) {
-		expr := current.SyntaxCar()
-		expandedExpr, err := childExpander.ExpandExpression(expr)
-		if err != nil {
-			return nil, werr.WrapForeignErrorf(err, "%s: failed to expand body expression", formName)
-		}
-		expandedExprs = append(expandedExprs, expandedExpr)
-		if isSyntaxFormWithKeyword(expandedExpr, "define") {
-			hasDefine = true
-		}
-		cdr := current.SyntaxCdr()
-		nextPair, ok := cdr.(*syntax.SyntaxPair)
-		if ok { //nolint:gocritic // ifElseChain: type assertion + value check, not a switch candidate
-			current = nextPair
-		} else if !syntax.IsSyntaxEmptyList(cdr) {
-			return nil, werr.WrapForeignErrorf(werr.ErrNotAList, "%s: body must be a proper list", formName)
-		} else {
-			break
-		}
-	}
-
-	// Build result: (begin body...) or ((lambda () (begin body...))) if has defines
-	beginSym := syntax.NewSyntaxSymbol("begin", sc)
-	beginBody := syntax.SyntaxList(sc, expandedExprs...)
-	beginExpr := syntax.NewSyntaxCons(beginSym, beginBody, sc)
-
-	if hasDefine {
-		// Wrap in lambda to create new runtime scope for defines
-		lambdaSym := syntax.NewSyntaxSymbol("lambda", sc)
-		emptyArgs := syntax.SyntaxEmptyList
-		lambdaExpr := syntax.SyntaxList(sc, lambdaSym, emptyArgs, beginExpr)
-		return syntax.SyntaxList(sc, lambdaExpr), nil
-	}
-
-	return beginExpr, nil
-}
-
-// isSyntaxFormWithKeyword checks if expr is a syntax pair whose car is
-// a syntax symbol with the given keyword.
-func isSyntaxFormWithKeyword(expr syntax.SyntaxValue, keyword string) bool {
-	pair, ok := expr.(*syntax.SyntaxPair)
-	if !ok || syntax.IsSyntaxEmptyList(pair) {
-		return false
-	}
-	sym, ok := pair.SyntaxCar().(*syntax.SyntaxSymbol)
-	if !ok {
-		return false
-	}
-	return sym.Sym.Key == keyword
-}
-
-// expandWithBindingScope implements the (with-binding-scope (id ...) body) form.
-//
-// This primitive expander creates a fresh "binding scope" and adds it to the entire
-// body. This is essential for hygienic macro expansion of binding forms like `let`.
-//
-// When a binding form (let, letrec, etc.) expands, it wraps its output in
-// with-binding-scope. The expander then:
-//  1. Creates a fresh scope S
-//  2. Adds S to the entire body (binding sites AND references)
-//  3. Returns the scoped body for further expansion
-//
-// This ensures that:
-//   - Each let form creates a unique scope
-//   - Binding identifiers and their references share that scope
-//   - Nested lets have different scopes, enabling hygiene
-//
-// Example:
-//
-//	(let ((x 1)) (+ x 1))
-//	→ macro expands to: (with-binding-scope (x) ((lambda (x) (+ x 1)) 1))
-//	→ expander adds scope S to body: ((lambda (x+S) (+ x+S 1)) 1)
-//	→ returns: ((lambda (x+S) (+ x+S 1)) 1)
-//
-// The identifier list (x) is currently unused but reserved for future use
-// (e.g., selective scope application or debugging).
-func (p *ExpanderTimeContinuation) expandWithBindingScope(_ *syntax.SyntaxSymbol, expr syntax.SyntaxValue) (syntax.SyntaxValue, error) {
-	// expr is the cdr of (with-binding-scope (id ...) body)
-	// which is ((id ...) body)
-	pair, ok := expr.(*syntax.SyntaxPair)
-	if !ok || syntax.IsSyntaxEmptyList(pair) {
-		return nil, werr.WrapForeignErrorf(werr.ErrInvalidSyntax, "with-binding-scope: expected (with-binding-scope (id ...) body)")
-	}
-
-	// Get the identifier list - these are the identifiers being bound
-	idListStx := pair.SyntaxCar()
-
-	// Get the body
-	cdr := pair.SyntaxCdr()
-	bodyPair, ok := cdr.(*syntax.SyntaxPair)
-	if !ok || syntax.IsSyntaxEmptyList(bodyPair) {
-		return nil, werr.WrapForeignErrorf(werr.ErrInvalidSyntax, "with-binding-scope: missing body")
-	}
-	body := bodyPair.SyntaxCar()
-
-	// Create a fresh binding scope
-	bindingScope := syntax.NewScopeWithLabel("binding")
-
-	// Add the scope to the entire body
-	// This adds the scope to ALL identifiers in the body, including:
-	// - Lambda parameters (binding sites)
-	// - References to those parameters in the lambda body
-	// - Any other identifiers
-	scopedBody := syntax.AddScopeToSyntax(body, bindingScope)
-
-	// Extract bound identifiers and create placeholder bindings for them.
-	// This is critical for R7RS §4.3.2 auxiliary syntax hygiene: when a macro
-	// like cond checks if => has been bound, it needs to find these placeholder
-	// bindings in the expand-time environment.
-	boundIds := extractIdentifierList(idListStx)
-	if len(boundIds) > 0 {
-		// Create a child expand environment with placeholder bindings
-		localExpandEnv := environment.NewLocalEnvironment(len(boundIds))
-		childExpandEnv := environment.NewEnvironmentFrameWithParent(localExpandEnv, p.env)
-
-		// Add placeholder bindings for each bound identifier.
-		// The scopes include the binding scope we just created.
-		for _, id := range boundIds {
-			// Get the identifier's current scopes and add the binding scope
-			idScopes := id.Scopes()
-			newScopes := make([]*syntax.Scope, len(idScopes)+1)
-			copy(newScopes, idScopes)
-			newScopes[len(idScopes)] = bindingScope
-
-			sym := id.Sym
-			childExpandEnv.MaybeCreateLocalBindingWithScopes(sym, environment.BindingTypeVariable, newScopes, id.SourceContext())
-		}
-
-		// Continue expansion with the child environment
-		childExpander := NewExpanderTimeContinuation(p.ctx, childExpandEnv)
-		return childExpander.ExpandExpression(scopedBody)
-	}
-
-	// No bound identifiers - just continue with current environment
-	return p.ExpandExpression(scopedBody)
-}
-
-// expandSyntaxError handles the (syntax-error message arg ...) form.
-// R7RS §4.3.1: syntax-error signals a compile-time error during macro expansion.
-// When encountered, it raises a compilation error with the given message and
-// arguments formatted as irritants.
-//
-// This allows macro authors to provide meaningful error messages for invalid uses:
-//
-//	(define-syntax must-be-pair
-//	  (syntax-rules ()
-//	    ((must-be-pair (a . b)) 'ok)
-//	    ((must-be-pair x) (syntax-error "expected a pair" x))))
-func (p *ExpanderTimeContinuation) expandSyntaxError(_ *syntax.SyntaxSymbol, expr syntax.SyntaxValue) (syntax.SyntaxValue, error) {
-	// Extract message (required first argument)
-	pair, ok := expr.(*syntax.SyntaxPair)
-	if !ok || syntax.IsSyntaxEmptyList(pair) {
-		return nil, werr.WrapForeignErrorf(werr.ErrInvalidSyntax, "syntax-error: missing message argument")
-	}
-
-	// Get the message
-	msgVal := pair.SyntaxCar()
-	var message string
-	switch m := msgVal.(type) {
-	case *syntax.SyntaxObject:
-		str, ok := m.Unwrap().(*values.String)
-		if ok {
-			message = str.Value
-		} else {
-			message = m.Unwrap().SchemeString()
-		}
-	case *syntax.SyntaxSymbol:
-		message = m.Sym.Key
-	default:
-		message = msgVal.SchemeString()
-	}
-
-	// Collect irritants (remaining arguments)
-	var irritants []string
-	rest := pair.SyntaxCdr()
-	for {
-		restPair, ok := rest.(*syntax.SyntaxPair)
-		if !ok || syntax.IsSyntaxEmptyList(restPair) {
-			break
-		}
-		irritant := restPair.SyntaxCar()
-		irritants = append(irritants, irritant.SchemeString())
-		rest = restPair.SyntaxCdr()
-	}
-
-	// Format the error message
-	if len(irritants) > 0 {
-		return nil, werr.WrapForeignErrorf(werr.ErrInvalidSyntax, "syntax-error: %s: %s", message, formatIrritants(irritants))
-	}
-	return nil, werr.WrapForeignErrorf(werr.ErrInvalidSyntax, "syntax-error: %s", message)
-}
-
-// formatIrritants joins irritants with commas for error display.
-func formatIrritants(irritants []string) string {
-	return strings.Join(irritants, ", ")
-}
-
-// expandBeginForm expands (begin expr ...) by expanding all subexpressions.
-// Uses ExpandBodyWithDefineSyntax to compile define-syntax forms immediately,
-// ensuring macros defined in begin are available to subsequent forms.
-func (p *ExpanderTimeContinuation) expandBeginForm(sym *syntax.SyntaxSymbol, expr syntax.SyntaxValue) (syntax.SyntaxValue, error) {
-	exprPair, ok := expr.(*syntax.SyntaxPair)
-	if ok && !syntax.IsSyntaxEmptyList(exprPair) {
-		// Collect forms from the begin body
-		forms, err := collectBodyExpressions(exprPair)
-		if err != nil {
-			return nil, werr.WrapForeignErrorf(err, "failed to collect begin body")
-		}
-
-		// Use ExpandBodyWithDefineSyntax to compile define-syntax forms immediately
-		// This ensures macros defined in begin are available to subsequent forms
-		expandedForms, err := p.ExpandBodyWithDefineSyntax(forms)
-		if err != nil {
-			return nil, werr.WrapForeignErrorf(err, "failed to expand begin body")
-		}
-
-		// Rebuild the begin form with expanded contents
-		expandedArgs := syntax.SyntaxEmptyList
-		for i := len(expandedForms) - 1; i >= 0; i-- {
-			expandedArgs = syntax.NewSyntaxCons(expandedForms[i], expandedArgs, sym.SourceContext())
-		}
-		return syntax.NewSyntaxCons(sym, expandedArgs, sym.SourceContext()), nil
-	}
-	return syntax.NewSyntaxCons(sym, expr, sym.SourceContext()), nil
-}
-
-// expandIfForm expands (if test consequent [alternative])
-func (p *ExpanderTimeContinuation) expandIfForm(sym *syntax.SyntaxSymbol, expr syntax.SyntaxValue) (syntax.SyntaxValue, error) {
-	pair, ok := expr.(*syntax.SyntaxPair)
-	if !ok || syntax.IsSyntaxEmptyList(pair) {
-		return syntax.NewSyntaxCons(sym, expr, sym.SourceContext()), nil
-	}
-
-	// Expand test
-	expandedTest, err := p.ExpandExpression(pair.SyntaxCar())
-	if err != nil {
-		return nil, werr.WrapForeignErrorf(err, "if: failed to expand test")
-	}
-
-	// Get consequent
-	cdrPair, ok := pair.SyntaxCdr().(*syntax.SyntaxPair)
-	if !ok || syntax.IsSyntaxEmptyList(cdrPair) {
-		return nil, werr.WrapForeignErrorf(werr.ErrInvalidSyntax, "if: missing consequent")
-	}
-
-	expandedConseq, err := p.ExpandExpression(cdrPair.SyntaxCar())
-	if err != nil {
-		return nil, werr.WrapForeignErrorf(err, "if: failed to expand consequent")
-	}
-
-	// Check for alternative
-	altPair, ok := cdrPair.SyntaxCdr().(*syntax.SyntaxPair)
-	if !ok || syntax.IsSyntaxEmptyList(altPair) {
-		// No alternative - (if test conseq)
-		args := syntax.SyntaxList(sym.SourceContext(), expandedTest, expandedConseq)
-		return syntax.NewSyntaxCons(sym, args, sym.SourceContext()), nil
-	}
-
-	// Expand alternative
-	expandedAlt, err := p.ExpandExpression(altPair.SyntaxCar())
-	if err != nil {
-		return nil, werr.WrapForeignErrorf(err, "if: failed to expand alternative")
-	}
-
-	// Build (if test conseq alt)
-	args := syntax.SyntaxList(sym.SourceContext(), expandedTest, expandedConseq, expandedAlt)
-	return syntax.NewSyntaxCons(sym, args, sym.SourceContext()), nil
-}
-
-// expandSetForm expands (set! var value)
-func (p *ExpanderTimeContinuation) expandSetForm(sym *syntax.SyntaxSymbol, expr syntax.SyntaxValue) (syntax.SyntaxValue, error) {
-	pair, ok := expr.(*syntax.SyntaxPair)
-	if !ok || syntax.IsSyntaxEmptyList(pair) {
-		return syntax.NewSyntaxCons(sym, expr, sym.SourceContext()), nil
-	}
-
-	// Keep variable unchanged
-	varExpr := pair.SyntaxCar()
-
-	// Expand value
-	cdrPair, ok := pair.SyntaxCdr().(*syntax.SyntaxPair)
-	if !ok || syntax.IsSyntaxEmptyList(cdrPair) {
-		return syntax.NewSyntaxCons(sym, expr, sym.SourceContext()), nil
-	}
-
-	expandedValue, err := p.ExpandExpression(cdrPair.SyntaxCar())
-	if err != nil {
-		return nil, werr.WrapForeignErrorf(err, "set!: failed to expand value")
-	}
-
-	// Build (set! var expanded-value)
-	args := syntax.SyntaxList(sym.SourceContext(), varExpr, expandedValue)
-	return syntax.NewSyntaxCons(sym, args, sym.SourceContext()), nil
-}
-
-// expandDefineForm expands (define var value) or (define (name . args) body ...)
-func (p *ExpanderTimeContinuation) expandDefineForm(sym *syntax.SyntaxSymbol, expr syntax.SyntaxValue) (syntax.SyntaxValue, error) {
-	pair, ok := expr.(*syntax.SyntaxPair)
-	if !ok || syntax.IsSyntaxEmptyList(pair) {
-		return syntax.NewSyntaxCons(sym, expr, sym.SourceContext()), nil
-	}
-
-	first := pair.SyntaxCar()
-	cdrPair, ok := pair.SyntaxCdr().(*syntax.SyntaxPair)
-	if !ok || syntax.IsSyntaxEmptyList(cdrPair) {
-		return syntax.NewSyntaxCons(sym, expr, sym.SourceContext()), nil
-	}
-
-	// Check if it's a function definition (define (name args...) body...)
-	_, isSymbol := first.(*syntax.SyntaxSymbol)
-	if !isSymbol {
-		// Function definition - first is (name args...)
-		// Expand the body expressions
-		expandedBody, err := p.ExpandSyntaxArgumentList(cdrPair)
-		if err != nil {
-			return nil, werr.WrapForeignErrorf(err, "define: failed to expand body")
-		}
-		args := syntax.NewSyntaxCons(first, expandedBody, sym.SourceContext())
-		return syntax.NewSyntaxCons(sym, args, sym.SourceContext()), nil
-	}
-
-	// Simple definition (define var value)
-	expandedValue, err := p.ExpandExpression(cdrPair.SyntaxCar())
-	if err != nil {
-		return nil, werr.WrapForeignErrorf(err, "define: failed to expand value")
-	}
-
-	// Build (define var expanded-value)
-	args := syntax.SyntaxList(sym.SourceContext(), first, expandedValue)
-	return syntax.NewSyntaxCons(sym, args, sym.SourceContext()), nil
-}
-
-// expandImportForm processes (import <import-set> ...) during expansion.
-// This loads libraries and makes their bindings (including macros) available
-// for subsequent forms. Without this, imported macros wouldn't be recognized
-// during expansion since imports were only processed during compilation.
-//
-// R7RS §5.2: Import declarations must be processed before expressions that
-// use the imported bindings. For macros, this means processing during expansion.
-func (p *ExpanderTimeContinuation) expandImportForm(sym *syntax.SyntaxSymbol, expr syntax.SyntaxValue) (syntax.SyntaxValue, error) {
-	// expr is the arguments after 'import': (<import-set> ...)
-	if syntax.IsSyntaxEmptyList(expr) {
-		// Empty import is valid, return unchanged
-		return syntax.NewSyntaxCons(sym, expr, sym.SourceContext()), nil
-	}
-
-	importSets, ok := expr.(*syntax.SyntaxPair)
-	if !ok {
-		return nil, werr.WrapForeignErrorf(werr.ErrInvalidSyntax, "import: expected list of import sets")
-	}
-
-	// Process each import set to load libraries and copy bindings
-	_, err := syntax.SyntaxForEach(p.ctx, importSets, func(ctx context.Context, _ int, _ bool, importSetExpr syntax.SyntaxValue) error {
-		return ResolveAndInstallImportSet(ctx, importSetExpr.UnwrapAll(), p.env, environment.PhaseExpand)
-	})
-
-	if err != nil {
-		return nil, err
-	}
-
-	// Return the import form unchanged for later compilation
-	return syntax.NewSyntaxCons(sym, expr, sym.SourceContext()), nil
-}
-
-// expandLambdaForm expands (lambda (args...) body...)
-//
-// R7RS §4.2.2: Lambda parameters shadow outer bindings including macros and
-// primitive forms. This function creates a child environment with the formals
-// as local variable bindings before expanding the body, ensuring that references
-// to parameter names (like `if`, `let`) don't get treated as special forms.
-//
-// R7RS §4.3.2: Auxiliary syntax hygiene. Lambda adds a scope to both formals
-// and body BEFORE expanding inner macros. This ensures that identifiers in the
-// body (like `=>` in a `cond`) carry the lambda's scope, enabling correct
-// `free-identifier=?` comparisons during macro pattern matching.
-//
-// R7RS §5.3: Internal define-syntax forms are processed before expanding the
-// rest of the body, so locally-defined macros are visible to subsequent forms.
-func (p *ExpanderTimeContinuation) expandLambdaForm(sym *syntax.SyntaxSymbol, expr syntax.SyntaxValue) (syntax.SyntaxValue, error) {
-	pair, ok := expr.(*syntax.SyntaxPair)
-	if !ok || syntax.IsSyntaxEmptyList(pair) {
-		return syntax.NewSyntaxCons(sym, expr, sym.SourceContext()), nil
-	}
-
-	formals := pair.SyntaxCar()
-	cdrVal := pair.SyntaxCdr()
-	cdrPair, ok := cdrVal.(*syntax.SyntaxPair)
-	if !ok || syntax.IsSyntaxEmptyList(cdrPair) {
-		return syntax.NewSyntaxCons(sym, expr, sym.SourceContext()), nil
-	}
-
-	// Create a scope for this lambda's bindings.
-	// This scope is added to both formals and body BEFORE any inner expansion,
-	// ensuring that pattern matching in inner macros (like cond) can correctly
-	// detect when identifiers (like =>) have been bound by this lambda.
-	//
-	// This also maintains the compiler's fast-path invariant: every symbol inside
-	// a local binding context has at least one scope (lambdaScope), so symbols
-	// with empty scopes can safely skip scope-aware resolution. See CompileSymbol.
-	lambdaScope := syntax.NewScopeWithLabel("lambda")
-
-	// Add lambda scope to formals and body
-	formalsStx := syntax.AddScopeToSyntax(formals, lambdaScope)
-	bodyWithScope := cdrPair.AddScope(lambdaScope).(*syntax.SyntaxPair)
-
-	// Extract formal parameter symbols (now with lambda scope included)
-	formalSyms := extractFormalSymbols(formalsStx)
-
-	// Create a child environment with the formals as local variable bindings.
-	// The bindings include the lambda scope, so lookups will find them.
-	childEnv := environment.NewEnvironmentFrameWithParent(
-		environment.NewLocalEnvironment(0),
-		p.env,
-	)
-	for _, fs := range formalSyms {
-		childEnv.MaybeCreateLocalBindingWithScopes(fs.sym, environment.BindingTypeVariable, fs.scopes, fs.source)
-	}
-
-	// R7RS §5.3: Process define-syntax forms before expanding subsequent expressions
-	// This makes locally-defined macros visible to later body expressions
-	bodyExprs, err := collectBodyExpressions(bodyWithScope)
-	if err != nil {
-		return nil, werr.WrapForeignErrorf(err, "lambda: invalid body expression")
-	}
-
-	// Handle the case where body is wrapped in (begin ...) - common from let macro
-	unwrappedExprs, wasBeginWrapped := unwrapBeginBodyWithFlag(bodyExprs)
-
-	// Expand body in the child environment, compiling define-syntax as encountered
-	childExpander := NewExpanderTimeContinuation(p.ctx, childEnv)
-	expandedExprs, err := childExpander.ExpandBodyWithDefineSyntax(unwrappedExprs)
-	if err != nil {
-		return nil, werr.WrapForeignErrorf(err, "lambda: failed to expand body")
-	}
-
-	// Rebuild the body as a syntax list
-	var expandedBody syntax.SyntaxValue
-	if wasBeginWrapped {
-		// Re-wrap in begin
-		beginSym := syntax.NewSyntaxSymbol("begin", sym.SourceContext())
-		innerList := syntax.SyntaxList(sym.SourceContext(), expandedExprs...)
-		beginForm := syntax.NewSyntaxCons(beginSym, innerList, sym.SourceContext())
-		expandedBody = syntax.SyntaxList(sym.SourceContext(), beginForm)
-	} else {
-		expandedBody = syntax.SyntaxList(sym.SourceContext(), expandedExprs...)
-	}
-
-	// Build (lambda formals expanded-body...)
-	args := syntax.NewSyntaxCons(formalsStx, expandedBody, sym.SourceContext())
-	return syntax.NewSyntaxCons(sym, args, sym.SourceContext()), nil
-}
-
-// collectBodyExpressions collects all expressions from a body syntax pair into a slice.
-func collectBodyExpressions(body *syntax.SyntaxPair) ([]syntax.SyntaxValue, error) {
-	var exprs []syntax.SyntaxValue
-	current := body
-	for !syntax.IsSyntaxEmptyList(current) {
-		exprs = append(exprs, current.SyntaxCar())
-		cdr := current.SyntaxCdr()
-		nextPair, ok := cdr.(*syntax.SyntaxPair)
-		if ok { //nolint:gocritic // ifElseChain: type assertion + value check, not a switch candidate
-			current = nextPair
-		} else if syntax.IsSyntaxEmptyList(cdr) {
-			break
-		} else {
-			return nil, werr.WrapForeignErrorf(werr.ErrNotAList, "body must be a proper list")
-		}
-	}
-	return exprs, nil
-}
-
-// unwrapBeginBodyWithFlag handles the case where a lambda body is a single (begin ...) form.
-// This is common from let macro expansion: (let () body...) -> ((lambda () (begin body...)))
-// Returns the contents of the begin and a flag indicating if unwrapping occurred.
-func unwrapBeginBodyWithFlag(exprs []syntax.SyntaxValue) ([]syntax.SyntaxValue, bool) {
-	if len(exprs) != 1 {
-		return exprs, false
-	}
-	pair, ok := exprs[0].(*syntax.SyntaxPair)
-	if !ok {
-		return exprs, false
-	}
-	carSym, ok := pair.SyntaxCar().(*syntax.SyntaxSymbol)
-	if !ok {
-		return exprs, false
-	}
-	sym, ok := carSym.Unwrap().(*values.Symbol)
-	if !ok || sym.Key != "begin" {
-		return exprs, false
-	}
-	// It's (begin ...), extract the contents
-	cdr := pair.Cdr()
-	cdrPair, ok := cdr.(*syntax.SyntaxPair)
-	if !ok {
-		return exprs, false
-	}
-	innerExprs, err := collectBodyExpressions(cdrPair)
-	if err != nil {
-		return exprs, false
-	}
-	return innerExprs, true
-}
-
-// formalSymbol pairs a symbol with its scopes for formal parameter tracking.
-type formalSymbol struct {
-	sym    *values.Symbol
-	scopes []*syntax.Scope
-	source *syntax.SourceContext
-}
-
-// extractFormalSymbols extracts symbols from a lambda formals expression.
-// Handles proper lists (x y z), improper lists (x y . rest), and single symbols (args).
-func extractFormalSymbols(formals syntax.SyntaxValue) []formalSymbol {
-	var result []formalSymbol
-
-	switch f := formals.(type) {
-	case *syntax.SyntaxSymbol:
-		// Rest argument: (lambda args body...)
-		result = append(result, formalSymbol{f.Sym, f.Scopes(), f.SourceContext()})
-	case *syntax.SyntaxPair:
-		// List of arguments: (lambda (x y z) body...) or (lambda (x y . rest) body...)
-		current := f
-		for !syntax.IsSyntaxEmptyList(current) {
-			car := current.SyntaxCar()
-			sym, ok := car.(*syntax.SyntaxSymbol)
-			if ok {
-				result = append(result, formalSymbol{sym.Sym, sym.Scopes(), sym.SourceContext()})
-			}
-			cdr := current.SyntaxCdr()
-			nextPair, ok := cdr.(*syntax.SyntaxPair)
-			if ok {
-				current = nextPair
-			} else {
-				sym, ok := cdr.(*syntax.SyntaxSymbol)
-				if ok {
-					// Improper list: (x y . rest)
-					result = append(result, formalSymbol{sym.Sym, sym.Scopes(), sym.SourceContext()})
-				}
-				break
-			}
-		}
-	}
-
-	return result
-}
-
-// extractIdentifierList extracts SyntaxSymbols from an identifier list.
-// This is used by with-binding-scope to get the bound identifiers.
-// Returns the symbols with their scopes preserved.
-func extractIdentifierList(idList syntax.SyntaxValue) []*syntax.SyntaxSymbol {
-	var result []*syntax.SyntaxSymbol
-
-	pair, ok := idList.(*syntax.SyntaxPair)
-	if !ok {
-		// Single identifier or empty
-		sym, ok := idList.(*syntax.SyntaxSymbol)
-		if ok {
-			return []*syntax.SyntaxSymbol{sym}
-		}
-		return nil
-	}
-
-	// List of identifiers
-	current := pair
-	for !syntax.IsSyntaxEmptyList(current) {
-		car := current.SyntaxCar()
-		sym, ok := car.(*syntax.SyntaxSymbol)
-		if ok {
-			result = append(result, sym)
-		}
-		cdr := current.SyntaxCdr()
-		nextPair, ok := cdr.(*syntax.SyntaxPair)
-		if ok {
-			current = nextPair
-		} else {
-			break
-		}
-	}
-
-	return result
-}
-
-// expandCaseLambdaForm expands (case-lambda (formals body...) ...)
-func (p *ExpanderTimeContinuation) expandCaseLambdaForm(sym *syntax.SyntaxSymbol, expr syntax.SyntaxValue) (syntax.SyntaxValue, error) {
-	pair, ok := expr.(*syntax.SyntaxPair)
-	if !ok || syntax.IsSyntaxEmptyList(pair) {
-		return syntax.NewSyntaxCons(sym, expr, sym.SourceContext()), nil
-	}
-
-	// Expand each clause
-	var expandedClauses []syntax.SyntaxValue
-	current := pair
-	for !syntax.IsSyntaxEmptyList(current) {
-		clauseVal := current.SyntaxCar()
-		clauseStx := clauseVal
-		// Each clause is (formals body...)
-		clausePair, ok := clauseStx.(*syntax.SyntaxPair)
-		if !ok || syntax.IsSyntaxEmptyList(clausePair) {
-			// Keep malformed clauses as-is, let validator report errors
-			expandedClauses = append(expandedClauses, clauseStx)
-		} else {
-			// Keep formals unchanged
-			formals := clausePair.SyntaxCar()
-			formalsStx := formals
-			// Expand body
-			cdrVal := clausePair.SyntaxCdr()
-			cdrPair, ok := cdrVal.(*syntax.SyntaxPair)
-			if !ok || syntax.IsSyntaxEmptyList(cdrPair) {
-				// No body - keep clause as-is
-				expandedClauses = append(expandedClauses, clauseStx)
-			} else {
-				expandedBody, err := p.ExpandSyntaxArgumentList(cdrPair)
-				if err != nil {
-					return nil, werr.WrapForeignErrorf(err, "case-lambda: failed to expand clause body")
-				}
-
-				// Build (formals expanded-body...)
-				expandedClause := syntax.NewSyntaxCons(formalsStx, expandedBody, clausePair.SourceContext())
-				expandedClauses = append(expandedClauses, expandedClause)
-			}
-		}
-
-		cdr := current.SyntaxCdr()
-		nextPair, ok := cdr.(*syntax.SyntaxPair)
-		if ok {
-			current = nextPair
-		} else {
-			break
-		}
-	}
-
-	// Build (case-lambda expanded-clauses...)
-	clauseList := syntax.SyntaxList(sym.SourceContext(), expandedClauses...)
-	return syntax.NewSyntaxCons(sym, clauseList, sym.SourceContext()), nil
 }
 
 // ExpandSyntaxExpression checks if sym is a macro and expands it, or returns
